added cropland classification split

+++ notebooks/cbutsko_utils.py
@@ -384,8 +384,8 @@ def process_raw_features_input_df(
         if tcol in tdf.columns:
             tdf[tcol] = pd.to_datetime(tdf[tcol])
 
-    tdf['cropland_wc'] = tdf['landcover_wc']==11
-    tdf['cropland_ec'] = tdf['landcover']==11
+    tdf['cropland_wc'] = (tdf['landcover_wc']==11).astype(int)
+    tdf['cropland_ec'] = (tdf['landcover']==11).astype(int)
 
     tdf.set_index(['sample_id'], inplace=True)
